refactor: split sqlite task event publication
